+++ src/layout/sidebar/menu/Menu.tsx
@@ -6,6 +6,7 @@ import { usePages } from 'src/routes'
 
 import type { MenuProps } from 'antd'
 import { Menu } from 'antd'
+import { LanguageToggle } from 'src/pages/EasterEgg/LanguageToggle'
 
 type MenuItem = Required<MenuProps>['items'][number]
 function getItem(
@@ -23,18 +24,12 @@ function getItem(
 }
 
 const MainMenu = () => {
-  const { t, i18n } = useTranslation()
+  const { t } = useTranslation()
   const pages = usePages()
   const items: MenuItem[] = pages.map((itm) => {
     return getItem(<Link to={t(itm.path)}>{t(itm.label)}</Link>, itm.path, itm.icon)
   })
-  const [currentLanguage, setCurrentLanguage] = useState('en')
 
-  const handleChangeLanguage = () => {
-    const newLanguage = currentLanguage === 'en' ? 'he' : 'en'
-    setCurrentLanguage(newLanguage)
-    i18n.changeLanguage(newLanguage)
-  }
   const location = useLocation()
   const [current, setCurrent] = useState(
     location.pathname === '/' || location.pathname === '' ? '/dashboard' : location.pathname,
@@ -60,7 +55,7 @@ const MainMenu = () => {
         mode="inline"
         items={items}
       />
-      {null && <button onClick={handleChangeLanguage}>Change Language</button>}
+      {<LanguageToggle />}
     </>
   )
 }
